requirements.txt
# Core ML & NLP
datasets>=2.18.0
huggingface-hub>=0.22.0
transformers>=4.39.0
torch>=2.2.0
scikit-learn>=1.4.0
scipy>=1.12.0
numpy>=1.26.0
pandas>=2.2.0

# NLP Utilities
langdetect>=1.0.9
langid>=1.1.6
fasttext-wheel>=0.9.2
lingua-language-detector>=2.0.2

# Unicode & Script Analysis
unicodedata2>=15.1.0
regex>=2023.12.25

# API & Serving
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
pydantic>=2.6.0
httpx>=0.27.0

# Visualization & Dashboard
gradio>=4.24.0
matplotlib>=3.8.0
seaborn>=0.13.2
plotly>=5.20.0
wordcloud>=1.9.3

# Experiment Tracking
mlflow>=2.11.0

# Utilities
tqdm>=4.66.0
rich>=13.7.0
loguru>=0.7.2
omegaconf>=2.3.0
pyyaml>=6.0.1
joblib>=1.3.0
click>=8.1.7
tabulate>=0.9.0

# Testing
pytest>=8.1.0
pytest-asyncio>=0.23.0
